move ytplayer to separate component, clean up
