+++ src/components/YoutubePlayer/YoutubePlayer.module.css
@@ -1,0 +1,99 @@
+.errorDisplay {
+    width: 480px;
+    height: 270px;
+    background-color: var(--surface-color);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    color: var(--text-secondary);
+    flex-direction: column;
+    gap: 10px;
+    padding: 1rem;
+
+    @media (max-width: 768px) {
+        width: 100%;
+        height: 250px;
+    }
+}
+
+.retryButton {
+    padding: 5px 10px;
+    background: var(--accent-color);
+    color: white;
+    border: none;
+    border-radius: 4px;
+}
+
+.thumbnailContainer {
+    width: 616px;
+    height: 346px;
+    position: relative;
+    background-color: var(--surface-color);
+    overflow: hidden;
+    border-radius: 8px;
+
+    @media (max-width: 768px) {
+        width: 100%;
+        height: 220px;
+    }
+
+    & img {
+        width: 100%;
+        height: 100%;
+        object-fit: cover;
+    }
+}
+
+.thumbnailOverlay {
+    position: absolute;
+    inset: 0;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background-color: rgba(0, 0, 0, 0.2);
+
+    /* Red button */
+    &>div {
+        width: 60px;
+        height: 60px;
+        border-radius: 50%;
+        background-color: rgba(255, 0, 0, 0.8);
+        display: flex;
+        align-items: center;
+        justify-content: center;
+
+        /* Triangle Shape */
+        &>div {
+            width: 0;
+            height: 0;
+            border-top: 15px solid transparent;
+            border-bottom: 15px solid transparent;
+            border-left: 24px solid white;
+            margin-left: 5px;
+        }
+
+        @media (max-width: 768px) {
+            width: 50px;
+            height: 50px;
+
+            &::after {
+                border-top: 12px solid transparent;
+                border-bottom: 12px solid transparent;
+                border-left: 20px solid white;
+            }
+        }
+    }
+}
+
+.videoContainer {
+    width: 616px;
+    height: 346px;
+    background-color: var(--surface-color);
+    overflow: hidden;
+}
+
+.iframeContainer {
+    width: 100%;
+    height: 100%;
+    border: none;
+}
